Test/inventory

+++ src/main/kotlin/io/github/dockyardmc/events/ServerTickEvent.kt
@@ -3,6 +3,6 @@ package io.github.dockyardmc.events
 import io.github.dockyardmc.annotations.EventDocumentation
 
 @EventDocumentation("when server ticks", false)
-class ServerTickEvent(val serverTicks: Int): Event {
+class ServerTickEvent(val serverTicks: Long): Event {
     override val context = Event.Context(isGlobalEvent = true)
 }
